Implement model loading, TODO: load textures

+++ nautilus/NautilusCameraFPS.cpp
@@ -39,8 +39,8 @@ void NautilusCameraFPS::cursor(GLFWwindow* _window, double _x, double _y) {
         lastY = _y;
         xOff *= this->m_sens;
         yOff *= this->m_sens;
-        this->m_angles.m_yaw += xOff;
-        this->m_angles.m_pitch += yOff;
+        this->m_angles.m_yaw += static_cast< float >(xOff);
+        this->m_angles.m_pitch += static_cast< float >(yOff);
         this->m_angles.m_pitch = std::clamp(this->m_angles.m_pitch, nautilus::defaults::CAMERA_MIN_PITCH, nautilus::defaults::CAMERA_MAX_PITCH);
         this->update();
     }
@@ -48,7 +48,7 @@ void NautilusCameraFPS::cursor(GLFWwindow* _window, double _x, double _y) {
 
 void NautilusCameraFPS::scroll(GLFWwindow* _window, double _dx, double _dy) {
     if(this->m_inputEnabled) {
-        this->m_fov -= _dy;
+        this->m_fov -= static_cast< float >(_dy);
         this->m_fov = std::clamp(this->m_fov, nautilus::defaults::CAMERA_MIN_FOV, nautilus::defaults::CAMERA_MAX_FOV);
     }
 }
